+++ src/main/java/ladder/domain/DirectionGenerator.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 public interface DirectionGenerator {
     boolean generate();
