+++ src/main/java/ladder/domain/RandomDirectionGenerator.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 import java.util.Random;
