+++ src/main/java/ladder/domain/Users.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 import util.CollectionUtil;
