+++ src/main/java/ladder/domain/UsersGenerator.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 import util.StringUtils;
