+++ src/test/java/ladder/domain/LadderGameTest.java
@@ -1,9 +1,10 @@
-package ladder;
+package ladder.domain;
 
+import ladder.domain.LadderGame;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-import static ladder.LadderGame.LADDER_HEIGHT_MUST_MORE_THEN_ONE;
+import static ladder.domain.LadderGame.LADDER_HEIGHT_MUST_MORE_THEN_ONE;
 import static org.assertj.core.api.Assertions.catchThrowable;
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
